refactor: move streaming selection attach into fallback_streams_select_streaming
Pure move: relocate the streaming fallback attach machinery
(_attach_selection_candidates_streaming with its nested _receive_next,
_attach_ready_selection_candidates, and their private helpers) verbatim into
a new sibling module fallback_streams_select_streaming.py. The names are
re-exported from fallback_streams_select so fallback_streams' import block and
all internal _fs.<name> call sites (and thus every existing patch target)
resolve unchanged. No behavior change.
